+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,6 +7,7 @@ development
 - IFI: Enable expression parsing with “Class” criteria in comfort form
 - For EP..A4 documents, display drawings of family members
 - Improve design and layout of query expression history chooser
+- Improve OPS connection error handling
 
 0.115.0
 =======